requirements.txt
tabulate==0.8.9
beautifulsoup4==4.10.0
bs4==0.0.1
python-dotenv==0.20.0
requests~=2.28.1
PyYAML~=6.0
setuptools~=57.0.0
aiogram~=2.22.1
aiosqlite~=0.17.0
pytest~=7.1.2
APScheduler~=3.6.3